Makefile.target: Introduce arch-obj
Add a new *obj- category, arch-obj. arch-obj-y will be a subset of the
existing obj-y. The difference, is arch-obj components are usable by
multi-arch builds. That is, following the single-arch target builds, the
already-built arch_obj components for multiple targets can be linked
together to form a multi-arch build that supports multi CPU archs.

Such a link is likely to have high numbers of namespace collisions. So
if the target arch supports multi-arch build, localise all symbols to a
single pre-linked object. The object does not need to export any APIs,
as all APIs are made available via QOM CPU hooks.

This forms a prerequisite for multi-arch support that there is no
definitions of symbols by arch-obj for use by core code.

target-foo is converted to arch-obj. But some CPUs may still need to
export APIs to device land (hw/). An example of this is the ARM
co-processor register interface. Such fns can be split off to new C
files in target-foo/hw dir where they remain obj-y for global
visibility. This creates a clearer separation of which
functions are system global and which are private to the CPU.

Signed-off-by: Peter Crosthwaite <crosthwaite.peter@gmail.com>

---CHANGELOG---
Changed since RFCv2
Convert all arches to arch-obj straight up.
Remove CONFIG_ARCH_SINGLE
---END---