include/exec: Split target_long def to new header
This is currently provided by cpu-defs and is a target specific
definition. However, to prepare for multi-arch only the bare minimum
content from cpu-defs.h should be exported to core code. And this is
all we need. So split it to a new header that the target_multi cpu.h
can include to save on having to include the ill-defined cpu-defs.h.

Allow multiple inclusion for multi-arch where multiple cpu.h's need
to be included and target_long will vary for each. This means that
target_[u]long needs to be changed from a typedef to a #define to
allow for redefinition.

Signed-off-by: Peter Crosthwaite <crosthwaite.peter@gmail.com>

---CHANGELOG---
Changed since RFC v2:
Convert target_[u]long to #define
---END---