target-*: Don't redefine cpu_exec()
This function needs to be converted to a QOM hook and virtualised for
multi-arch. This rename interferes, as cpu-qom will not have access
to the renaming causing name divergence. This rename doesn't really do
anything anyway so just delete it.

Cc: Blue Swirl <blauwirbel@gmail.com>
Cc: Riku Voipio <riku.voipio@iki.fi>
Signed-off-by: Peter Crosthwaite <crosthwaite.peter@gmail.com>

---CHANGELOG---
Changed since RFCv3:
Add bsd-user changes
Add change to target-tilegx (new arch)
Fix reference to "cpu_arm_exec" in comment
---END---