error: Factor out common error setter logic
Currently set_error, set_error_errno, set_error_win32 do pretty much
the same thing with only a little bit of string manipulation in the
error code cases. Coreify that string manipulation as a callback and
factor out the bulk of the error_set fns to a common helper that calls
the callback at the right time.

Signed-off-by: Peter Crosthwaite <crosthwaite.peter@gmail.com>