error: Add support for multiple errors
Allow errors to stack. If an error is already set, don't assert,
instead, form a linked list. Recent errors are at the front of the
list, older ones at the back.

The assertion against the destination erro already being set is
removed.

copy/free are all to call their functionality recursively.

Propagation is implemented as concatenation of two error lists.

Signed-off-by: Peter Crosthwaite <crosthwaite.peter@gmail.com>