cpus: Listify cpu_list() function
Multiple multi-arch targets may wish to implement cpu_list(). When the
command is called each should be called one after other. Create a list
to allow registrations. When cpu_list() is called the list is iterated
calling all implementors.

The original singleton #define'able cpu_list() mechanism remains in
place to allow gradual conversion of the target arches to the new way.

cpu_list_add() needs to visible from target-foo/*.c which means it
cannot be placed in cpus.c (common-obj). So put it in
cpu-exec-common.c. For consistency, move cpu_list() alongside it.

Signed-off-by: Peter Crosthwaite <crosthwaite.peter@gmail.com>

---CHANGELOG---
---END---