COVER: error: Add automatic error concatenation
---CHANGELOG---
Hi Markus and all,

This patch series adds support for automatically concatenating multiple
errors to the one Error *.

I'll start with what I am actually trying to do, which is get rid of the
boilerplate:

some_api_call(... , &local_err);
if (local_err) {
    error_propagate(errp, local_err);
    return;
}
some_similar_api_call(... , &local_err);
if (local_err) {
    error_propagate(errp, local_err);
    return;
}
some_similar_api_call(... , &local_err);
if (local_err) {
    error_propagate(errp, local_err);
    return;
}
...

It is usually 1 LOC for the API and 4 LOC for the error handling boiler
plate making our code hard reading. This is particularly bad in hw/arm
where we have a good number of fully QOMified SoCs and machine models,
each of which need these checks on recursive realize calls and friends.

The removed LOC just in ARM pays for the extra core code needed to
implement this. And the number of ARM machines is only going to grow.

So the plan is:

1: Allow an Error * to contain more that one actual error from API
calls.
2: Refactor key APIs (some_similar_api_call() in the above example)
to not fatal when previous errors have occured in dependencies.

Point 1 kind of got big on me. Patch 4 is the main event, listifying
errors. The follow up issue however, is it tricky to get a sane
definition of error_get_pretty for a multi-error. So instead the
strategy is to remove error_get_pretty() and replace with some error
API helper with well defined behaviour for multi-error. The two leading
uses of error get pretty are prefixing an error, and reporting an error
via a custom printf handler. So two new APIs are added for that (P5-6).
There aren't many error_get_pretty's left after that, and they
shouldn't be in the path of any multi-errors.

Point 2 is less about error API and more about machine generation.
Sysbus, Memory and Qdev all have APIs that depend on successful device-
init and realize calls for argument devices. As we are trying to remove
the error detection for those argument devs, those APIs need to tweaked
to handle realize failure. This actually wasn't as bad as I thought it
would be. See patches (7-9).

All patches after that walk the various major subsystems converting
error APIs to this new scheme and deleting now-unneeded boiler plate.
ARM is first (P10-15) seeing good clean up of propagate handers.

So the net result for these ARM machines, is error behaviour that is
something like a compiler. If any one thing fails, then machine-init
(compilation) fails. But an early fail does not stop machine-init
(compilation), instead it proceeds to then end collecting subsequent
errors as it goes.

Some other interesting food for thought is the qemu_fdt APIs which I
have been wanting to convert to error API but the local_err propagation
is going to be brutal in heavy users like e500.c. This would solve that
as fdt API could be easily made multi-error safe and clients like e500
can just collect multi-errors and single-fail at the end.

Long term, we can use this to catch cases of multiple genuine machine
init errors in the one boot but that is a secondary goal to simply
cutting down on code boilerplate. The best feature of this series is
the diffstat.

Patches 1-3 are cleanup that can be taken independent of the series.

Regards,
Peter
--END---

Signed-off-by: Peter Crosthwaite <crosthwaite.peter@gmail.com>