misc: imx31_ccm: Arrayify regsiters
Convert the registers to an array. This allows VMSD, reset and the
MMIO handlers to deal with the registers en-masse rather than
individually.

The behaviour where the LTR1 register is not writable is preserved.
According to the TRM, some fields of this register are writable but
others are not, so err on the side of caution by avoiding functional
change.

Signed-off-by: Peter Crosthwaite <crosthwaite.peter@gmail.com>
---
We could add the new register array to VMSD additively, to avoid the
version bump, but last I knew ARM still accepted version bumps. So
clean up while we still can.