arm: cpu: handle BE32 user-mode as BE
From a system emulation point of view, BE32 is best modelled as little
endian with address manipulations on subword accesses (to give the
illusion of BE). But user-mode cannot tell the difference and is
already implemented as straight BE. So handle the difference in the
endianess query, where USER mode is BE and system is not.

---CHANGELOG---
Changed since v1:
Rewrote commit subject and message formerly:
    arm: linux-user: don't set CPSR.E in BE32 mode
---END---

Signed-off-by: Peter Crosthwaite <crosthwaite.peter@gmail.com>