loader: add API to load elf header
Add an API to load an elf header header from a file. Populates a
buffer with the header contents, as well as a boolean for whether the
elf is 64b or not. Both arguments are optional.

---CHANGELOG---
Changed since v1 (PMM review):
Add filname to error messages
Remove unneeded lseek()
Add doc comment
---END---

Signed-off-by: Peter Crosthwaite <crosthwaite.peter@gmail.com>